On branch master
Initial commit - Sequential fetching for a given chapter
=========================================================

This is a proof-of-concept and many important features
have not been implemented. We need to loop over the courses,
and the chapters. And we need to create a folder structure to
save the videos.

To login, simply change 'user.info.example' to your own needs,
and save it as 'user.info'.

 Changes to be committed:
	new file:   the_jiu_jitsu_lifestyle_download.py
	new file:   user.info.example